requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-jose[cryptography]
passlib[argon2]
requests
psycopg2-binary
sqlalchemy
google-genai
pytest
pytest-mock


